requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
dill
scipy
pyaml
mlflow
dagshub
fastapi
uvicorn




#-e . if you want to uncomment and run it will automatically trigger the setup.py and the entire package for this project will be created.